* commit '8093a973765786817870c3ffe3eae4d8df691b63':
  ALG-5157 - Sonar rule.

+++ app/gtnForecasting/src/main/java/com/stpl/app/utils/QueryUtils.java
@@ -809,7 +809,7 @@ public class QueryUtils {
             for (Object temp : input) {
                 sql.replace(sql.indexOf("?"), sql.indexOf("?") + 1, String.valueOf(temp));
             }
-            int count = (Integer) HelperTableLocalServiceUtil.executeUpdateQueryCount(sql.toString());
+            int count = HelperTableLocalServiceUtil.executeUpdateQueryCount(sql.toString());
             return count > 0 ? Boolean.TRUE : Boolean.FALSE;
         } catch (Exception ex) {
             LOGGER.error(ex.getMessage());
@@ -828,7 +828,7 @@ public class QueryUtils {
                 }
             }
         }
-        int count = (Integer) HelperTableLocalServiceUtil.executeUpdateQueryCount(queryString.toString());
+        int count =  HelperTableLocalServiceUtil.executeUpdateQueryCount(queryString.toString());
         return count > 0 ? Boolean.TRUE : Boolean.FALSE;
     }
     
@@ -869,7 +869,7 @@ public class QueryUtils {
             for (Object temp : input) {
                 sql.replace(sql.indexOf(StringConstants.QUESTION.getConstant()), sql.indexOf(StringConstants.QUESTION.getConstant()) + 1, String.valueOf(temp));
             }
-            int count = (Integer) HelperTableLocalServiceUtil.executeUpdateQueryCount(QueryUtil.replaceTableNames(sql.toString(), session.getCurrentTableNames()));
+            int count = HelperTableLocalServiceUtil.executeUpdateQueryCount(QueryUtil.replaceTableNames(sql.toString(), session.getCurrentTableNames()));
             return count > 0 ? Boolean.TRUE : Boolean.FALSE;
         } catch (Exception ex) {
             LOGGER.error(ex.getMessage());
